Silence a couple of warnings by clippy (#247)
needless_lifetimes:

Clippy suggests that the lifetime is not needed and suggests eliding it.
While that is true, making this implicit would hide the fact that the
returned value borrows. While some may think this is obvious, I don't
think this is obvious for everyone.

If I could I would change the default to the opposite.

upper_case_acronyms:

While clippy is right here, this is how the instructions were named
historically. That said, I don't mind renaming it just think that it's
not worth it.